refactor to remove redundancy and improve coupling:
- use existing `evaluateVariable` func to do env var replacement
- get rid of potential unrelated functionality dumping ground (utils.go), giving that file a tighter definition and moving dir check to fsutils where other similar functionality already exists
