--- pkg/dependency/parser/java/pom/utils.go
@@ -1,21 +1,0 @@
-package pom
-
-import (
-	"os"
-	"strings"
-)
-
-func isDirectory(path string) (bool, error) {
-	fileInfo, err := os.Stat(path)
-	if err != nil {
-		return false, err
-	}
-	return fileInfo.IsDir(), err
-}
-
-func isProperty(version string) bool {
-	if version != "" && strings.HasPrefix(version, "${") && strings.HasSuffix(version, "}") {
-		return true
-	}
-	return false
-}
