+++ pkg/utils/fsutils/fs.go
@@ -58,8 +58,16 @@ func CopyFile(src, dst string) (int64, error) {
 }
 
 func DirExists(path string) bool {
-	f, err := os.Stat(path)
-	return err == nil && f.IsDir()
+	isDir, _ := CheckDirectory(path)
+	return isDir
+}
+
+func CheckDirectory(path string) (bool, error) {
+	fileInfo, err := os.Stat(path)
+	if err != nil {
+		return false, err
+	}
+	return fileInfo.IsDir(), err
 }
 
 func FileExists(filename string) bool {
